Manage and change the stars' states on click
We click on a star and it changes the states of
stars in the object until we reached the key we
just clicked on. Everything else will be the default
state.

That way we can always update our rating later on,
in case we change our mind.

We also want to update the shown score, so we use a new
variable for that, which is defaulted to null.