Create hover effect
Granted, I panicked here a bit.
I realized that I should've changed the visual
changes with classes and background images/image masks,
instead of changing the content. That makes hovering a lot
more annoying to deal with.

Hindsight is 20/20.

I could also have merged this into the setState function,
with optional parameters for the ref, what it's supposed to do,
and if the state change should update the score text or not.

The restore function resets everything back to the default color.
Also something that could've been done with classes.

I'm running out of time. D: